# optimize Java Map initialization