I fixed the Event initialize for the Java Map parameter case. the initialize method cannot receive a pure Java object parameter but only a proxied Java object and the only one we support is actually the `MapJavaProxy` and we can efficiently get to the original Map object instead of going through the hoops of `to_java`.