@colinsurprenant 2016-05-19
Benchmarks - *Ruby* is initializing the Event with a Ruby Hash parameter and *Java* is initializing the Event with a Java HashMap parameter

#### With PR

```
Warming up --------------------------------------
                Ruby     9.791k i/100ms
                Java    62.671k i/100ms
Calculating -------------------------------------
                Ruby    124.201k (± 8.7%) i/s -      1.234M
                Java      2.140M (±12.2%) i/s -     20.995M

Comparison:
                Java:  2139968.2 i/s
                Ruby:   124201.5 i/s - 17.23x slower
```

#### Master

```
Warming up --------------------------------------
                Ruby     9.396k i/100ms
                Java    45.526k i/100ms
Calculating -------------------------------------
                Ruby    117.353k (± 8.4%) i/s -      1.165M
                Java      1.187M (±11.8%) i/s -     11.609M

Comparison:
                Java:  1187122.0 i/s
                Ruby:   117352.6 i/s - 10.12x slower
```