@colinsurprenant 2016-05-25
rebased, merged in master and 5.0